Added 'about this text' link to all texts

+++ src/prefs.ts
@@ -354,7 +354,7 @@ export default class Prefs {
             data: null,
           },
         };
-        if (!this.stores[aStore].store.exists()) {
+        if (!useDefaultStore && !this.stores[aStore].store.exists()) {
           this.stores[aStore].store.writeFile(JSON_stringify({}));
         }
       } else throw new Error('Prefs has not been initialized!');
